+++ src/main/java/com/goit/popov/restaurant/dao/impl/UnitDAOImplJPA.java
@@ -5,7 +5,6 @@ import com.goit.popov.restaurant.model.Unit;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 
 /**
@@ -23,7 +22,7 @@ public class UnitDAOImplJPA implements UnitDAO {
         }
 
         @Override
-        public Unit getById(int id) {
+        public Unit getById(Long id) {
                 return sessionFactory.getCurrentSession().get(Unit.class, id);
         }
 }
